@@ -1,42 +1,42 @@
 module backend
 
-go 1.22rc1
+go 1.23
 
 require (
+	github.com/MatusOllah/slogcolor v1.4.0
 	github.com/alecthomas/template v0.0.0-20190718012654-fb15b899a751
 	github.com/dimiro1/banner v1.1.0
-	github.com/dusted-go/logging v1.1.3
-	github.com/ethereum/go-ethereum v1.13.8
-	github.com/felixge/httpsnoop v1.0.3
-	github.com/flatfeestack/go-lib v0.0.6
-	github.com/go-jose/go-jose/v3 v3.0.0
-	github.com/google/uuid v1.3.1
+	github.com/ethereum/go-ethereum v1.14.11
+	github.com/fatih/color v1.17.0
+	github.com/felixge/httpsnoop v1.0.4
+	github.com/go-jose/go-jose/v3 v3.0.3
+	github.com/google/uuid v1.6.0
 	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
-	github.com/mattn/go-sqlite3 v1.14.17
-	github.com/nspcc-dev/neo-go v0.105.0
+	github.com/mattn/go-sqlite3 v1.14.24
+	github.com/nspcc-dev/neo-go v0.106.3
+	github.com/sendgrid/sendgrid-go v3.16.0+incompatible
 	github.com/spaolacci/murmur3 v1.1.0
-	github.com/stretchr/testify v1.8.4
-	github.com/stripe/stripe-go/v76 v76.11.0
-	golang.org/x/text v0.14.0
+	github.com/stretchr/testify v1.9.0
+	github.com/stripe/stripe-go/v76 v76.25.0
+	golang.org/x/text v0.19.0
 )
 
 require (
-	github.com/btcsuite/btcd/btcec/v2 v2.2.0 // indirect
+	github.com/btcsuite/btcd/btcec/v2 v2.3.4 // indirect
 	github.com/common-nighthawk/go-figure v0.0.0-20200609044655-c4b36f998cf2 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.0.1 // indirect
-	github.com/hashicorp/golang-lru v0.6.0 // indirect
-	github.com/holiman/uint256 v1.2.4 // indirect
-	github.com/mattn/go-isatty v0.0.19 // indirect
+	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.2.0 // indirect
+	github.com/hashicorp/golang-lru/v2 v2.0.7 // indirect
+	github.com/holiman/uint256 v1.3.1 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mr-tron/base58 v1.2.0 // indirect
-	github.com/nspcc-dev/rfc6979 v0.2.0 // indirect
+	github.com/nspcc-dev/rfc6979 v0.2.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/rogpeppe/go-internal v1.10.0 // indirect
-	github.com/sendgrid/sendgrid-go v3.12.0+incompatible // indirect
-	github.com/sirupsen/logrus v1.9.2 // indirect
-	golang.org/x/crypto v0.17.0 // indirect
-	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/crypto v0.22.0 // indirect
+	golang.org/x/sys v0.22.0 // indirect
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
