@@ -1,41 +1,30 @@
 module fastauth
 
-go 1.20
+go 1.23
 
 require (
+	github.com/MatusOllah/slogcolor v1.4.0
 	github.com/dimiro1/banner v1.1.0
-	github.com/felixge/httpsnoop v1.0.3
-	github.com/flatfeestack/go-lib v0.0.6
-	github.com/go-jose/go-jose/v3 v3.0.0
-	github.com/gorilla/mux v1.8.0
-	github.com/gorilla/schema v1.2.0
+	github.com/fatih/color v1.17.0
+	github.com/felixge/httpsnoop v1.0.4
+	github.com/go-jose/go-jose/v3 v3.0.3
+	github.com/gorilla/mux v1.8.1
 	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
-	github.com/mattn/go-sqlite3 v1.14.19
-	github.com/prometheus/client_golang v1.15.1
-	github.com/sirupsen/logrus v1.9.3
-	github.com/stretchr/testify v1.8.4
+	github.com/mattn/go-sqlite3 v1.14.24
+	github.com/sendgrid/sendgrid-go v3.16.0+incompatible
+	github.com/stretchr/testify v1.9.0
 	github.com/xlzd/gotp v0.1.0
-	golang.org/x/crypto v0.17.0
-	golang.org/x/text v0.14.0
+	golang.org/x/crypto v0.28.0
+	golang.org/x/text v0.19.0
 )
 
 require (
-	github.com/beorn7/perks v1.0.1 // indirect
-	github.com/cespare/xxhash/v2 v2.2.0 // indirect
 	github.com/common-nighthawk/go-figure v0.0.0-20200609044655-c4b36f998cf2 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/kr/text v0.2.0 // indirect
-	github.com/mattn/go-isatty v0.0.12 // indirect
-	github.com/matttproud/golang_protobuf_extensions v1.0.4 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/prometheus/client_model v0.3.0 // indirect
-	github.com/prometheus/common v0.42.0 // indirect
-	github.com/prometheus/procfs v0.9.0 // indirect
-	github.com/rogpeppe/go-internal v1.9.0 // indirect
-	github.com/sendgrid/sendgrid-go v3.12.0+incompatible // indirect
-	golang.org/x/sys v0.15.0 // indirect
-	google.golang.org/protobuf v1.30.0 // indirect
+	golang.org/x/sys v0.26.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
