@@ -5,10 +5,16 @@ import (
 	"encoding/base32"
 	"encoding/json"
 	"fmt"
-	dbLib "github.com/flatfeestack/go-lib/database"
-	log "github.com/sirupsen/logrus"
+	"io"
+	"log/slog"
+	"os"
+	"regexp"
 	"strings"
 	"time"
+)
+
+var (
+	DB *sql.DB
 )
 
 type dbRes struct {
@@ -16,8 +22,6 @@ type dbRes struct {
 	password     []byte
 	refreshToken string
 	emailToken   *string
-	sms          *string
-	smsVerified  *time.Time
 	totp         *string
 	totpVerified *time.Time
 	errorCount   int
@@ -28,12 +32,12 @@ type dbRes struct {
 func findAuthByEmail(email string) (*dbRes, error) {
 	var res dbRes
 	var pw string
-	query := `SELECT sms, password, meta_system, refresh_token, 
-       				 email_token, totp, sms_verified, totp_verified, error_count, flow_type
+	query := `SELECT password, meta_system, refresh_token, 
+       				 email_token, totp, totp_verified, error_count, flow_type
               FROM auth WHERE email = $1`
-	err := dbLib.DB.QueryRow(query, email).Scan(
-		&res.sms, &pw, &res.metaSystem, &res.refreshToken, &res.emailToken,
-		&res.totp, &res.smsVerified, &res.totpVerified, &res.errorCount, &res.flowType)
+	err := DB.QueryRow(query, email).Scan(
+		&pw, &res.metaSystem, &res.refreshToken, &res.emailToken,
+		&res.totp, &res.totpVerified, &res.errorCount, &res.flowType)
 
 	if err != nil {
 		return nil, err
@@ -52,7 +56,7 @@ func insertUser(email string, pwRaw []byte, emailToken string, refreshToken stri
 		s1 := base32.StdEncoding.EncodeToString(pwRaw)
 		pw = &s1
 	}
-	stmt, err := dbLib.DB.Prepare(`INSERT INTO auth as a (email, password, email_token, refresh_token, flow_type, created_at) 
+	stmt, err := DB.Prepare(`INSERT INTO auth as a (email, password, email_token, refresh_token, flow_type, created_at) 
 								   VALUES ($1, $2, $3, $4, $5, $6)
 								   ON CONFLICT (email) DO
 								     UPDATE SET password=$2, email_token = $3
@@ -60,165 +64,147 @@ func insertUser(email string, pwRaw []byte, emailToken string, refreshToken stri
 	if err != nil {
 		return fmt.Errorf("prepare INSERT INTO auth for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(email, pw, emailToken, refreshToken, flowType, now)
-	return handleErr(res, err, "INSERT INTO auth", email)
+	return handleErrEffect(res, err, "error: INSERT INTO auth", email)
 }
 
 func deleteDbUser(email string) error {
-	stmt, err := dbLib.DB.Prepare(`DELETE FROM auth where email = $1`)
+	stmt, err := DB.Prepare(`DELETE FROM auth where email = $1`)
 	if err != nil {
 		return fmt.Errorf("prepare INSERT INTO auth for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(email, email)
-	return handleErr(res, err, "DELETE FROM auth", email)
+	return handleErrEffect(res, err, "error: DELETE FROM auth", email)
 }
 
 func updateRefreshToken(oldRefreshToken string, newRefreshToken string) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE auth SET refresh_token = $1 WHERE refresh_token = $2")
+	stmt, err := DB.Prepare("UPDATE auth SET refresh_token = $1 WHERE refresh_token = $2")
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE refreshTokenfor statement failed: %v", err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(newRefreshToken, oldRefreshToken)
-	return handleErrEffect(res, err, "UPDATE refreshToken", "n/a", false)
+	return handleErrNoEffect(res, err, "error: UPDATE refreshToken", "n/a")
 }
 
 func updateSystemMeta(email string, systemMeta string) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE auth SET meta_system = $1 WHERE email=$2")
+	stmt, err := DB.Prepare("UPDATE auth SET meta_system = $1 WHERE email=$2")
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE meta_system statement failed: %v", err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(systemMeta, email)
-	return handleErr(res, err, "UPDATE meta_system", "n/a")
+	return handleErrEffect(res, err, "error: UPDATE meta_system", "n/a")
 }
 
 func updatePasswordInvite(email string, emailToken string, newPw []byte) error {
 	pw := base32.StdEncoding.EncodeToString(newPw)
-	stmt, err := dbLib.DB.Prepare(`UPDATE auth SET password = $1, email_token = NULL 
+	stmt, err := DB.Prepare(`UPDATE auth SET password = $1, email_token = NULL 
 								   WHERE email = $2 AND email_token = $3 AND password IS NULL`)
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE auth password for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(pw, email, emailToken)
-	return handleErr(res, err, "UPDATE auth password invite", email)
+	return handleErrEffect(res, err, "error: UPDATE auth password invite", email)
 }
 
 func updatePasswordForgot(email string, forgetEmailToken string, newPw []byte) error {
 	pw := base32.StdEncoding.EncodeToString(newPw)
-	stmt, err := dbLib.DB.Prepare(`UPDATE auth SET password = $1, totp = NULL, sms = NULL, forget_email_token = NULL 
+	stmt, err := DB.Prepare(`UPDATE auth SET password = $1, totp = NULL, forget_email_token = NULL 
 								   WHERE email = $2 AND forget_email_token = $3`)
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE auth password for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(pw, email, forgetEmailToken)
-	return handleErr(res, err, "UPDATE auth password", email)
+	return handleErrEffect(res, err, "error: UPDATE auth password", email)
 }
 
 func updateEmailForgotToken(email string, token string) error {
 	//TODO: don't accept too old forget tokens
-	stmt, err := dbLib.DB.Prepare("UPDATE auth SET forget_email_token = $1 WHERE email = $2")
+	stmt, err := DB.Prepare("UPDATE auth SET forget_email_token = $1 WHERE email = $2")
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE auth forgetEmailToken for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(token, email)
-	return handleErr(res, err, "UPDATE auth forgetEmailToken", email)
+	return handleErrEffect(res, err, "error: UPDATE auth forgetEmailToken", email)
 }
 
 func updateTOTP(email string, totp string) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE auth SET totp = $1 WHERE email = $2 and totp IS NULL")
+	stmt, err := DB.Prepare("UPDATE auth SET totp = $1 WHERE email = $2 and totp IS NULL")
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE auth totp for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(totp, email)
-	return handleErr(res, err, "UPDATE auth totp", email)
-}
-
-func updateSMS(email string, totp string, sms string) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE auth SET totp = $1, sms = $2 WHERE email = $3 AND sms_verified IS NULL")
-	if err != nil {
-		return fmt.Errorf("prepare UPDATE auth totp for %v statement failed: %v", email, err)
-	}
-	defer dbLib.CloseAndLog(stmt)
-
-	res, err := stmt.Exec(totp, sms, email)
-	return handleErr(res, err, "UPDATE auth totp", email)
+	return handleErrEffect(res, err, "error: UPDATE auth totp", email)
 }
 
 func updateEmailToken(email string, token string) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE auth SET email_token = NULL WHERE email = $1 AND email_token = $2")
+	stmt, err := DB.Prepare("UPDATE auth SET email_token = NULL WHERE email = $1 AND email_token = $2")
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE auth for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(email, token)
-	return handleErr(res, err, "UPDATE auth", email)
-}
-
-func updateSMSVerified(email string, now time.Time) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE auth SET sms_verified = $1 WHERE email = $2 AND sms IS NOT NULL")
-	if err != nil {
-		return fmt.Errorf("prepare UPDATE auth for %v statement failed: %v", email, err)
-	}
-	defer dbLib.CloseAndLog(stmt)
-
-	res, err := stmt.Exec(now, email)
-	return handleErr(res, err, "UPDATE auth SMS timestamp", email)
+	return handleErrEffect(res, err, "error: UPDATE auth", email)
 }
 
 func updateTOTPVerified(email string, now time.Time) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE auth SET totp_verified = $1 WHERE email = $2 AND totp IS NOT NULL")
+	stmt, err := DB.Prepare("UPDATE auth SET totp_verified = $1 WHERE email = $2 AND totp IS NOT NULL")
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE auth for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(now, email)
-	return handleErr(res, err, "UPDATE auth totp timestamp", email)
+	return handleErrEffect(res, err, "error: UPDATE auth totp timestamp", email)
 }
 
 func incErrorCount(email string) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE auth set error_count = error_count + 1 WHERE email = $1")
+	stmt, err := DB.Prepare("UPDATE auth set error_count = error_count + 1 WHERE email = $1")
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE auth status for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(email)
-	return handleErr(res, err, "UPDATE auth errorCount", email)
+	return handleErrEffect(res, err, "error: UPDATE auth errorCount", email)
 }
 
 func resetCount(email string) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE auth set error_count = 0 WHERE email = $1")
+	stmt, err := DB.Prepare("UPDATE auth set error_count = 0 WHERE email = $1")
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE auth status for %v statement failed: %v", email, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(email)
-	return handleErr(res, err, "UPDATE auth status", email)
+	return handleErrEffect(res, err, "error: UPDATE auth status", email)
 }
 
-func handleErr(res sql.Result, err error, info string, email string) error {
-	return handleErrEffect(res, err, info, email, true)
+func handleErrEffect(res sql.Result, err error, info string, email string) error {
+	return handleErr(res, err, info, email, true)
 }
 
-func handleErrEffect(res sql.Result, err error, info string, email string, mustHaveEffect bool) error {
+func handleErrNoEffect(res sql.Result, err error, info string, email string) error {
+	return handleErr(res, err, info, email, false)
+}
+
+func handleErr(res sql.Result, err error, info string, email string, mustHaveEffect bool) error {
 	if err != nil {
 		return fmt.Errorf("%v query %v failed: %v", info, email, err)
 	}
@@ -277,13 +263,87 @@ func addInitialUsers() {
 			if len(userPwMeta) >= 2 {
 				err := addInitialUserWithMeta(userPwMeta[0], userPwMeta[1], metaSystem)
 				if err == nil {
-					log.Printf("insterted user %v", userPwMeta[0])
+					slog.Debug("insterted user %v", slog.String("userPwMeta[0]", userPwMeta[0]))
 				} else {
-					log.Printf("could not insert %v: %v", userPwMeta[0], err)
+					slog.Warn("could not insert %v: %v", slog.String("userPwMeta[0]", userPwMeta[0]), slog.Any("error", err))
 				}
 			} else {
-				log.Printf("username and password need to be seperated by ':'")
+				slog.Warn("username and password need to be seperated by ':'")
 			}
 		}
+	}
+}
+
+func CloseDb() {
+	err := DB.Close()
+	slog.Warn("could not close the db", slog.Any("error", err))
+}
+
+func InitDb(dbDriver string, dbPath string, dbScripts string) error {
+	// Open the connection
+	var err error
+	DB, err = sql.Open(dbDriver, dbPath)
+	if err != nil {
+		return err
+	}
+
+	//we wait for ten seconds to connect
+	err = DB.Ping()
+	now := time.Now()
+	for err != nil && now.Add(time.Duration(10)*time.Second).After(time.Now()) {
+		time.Sleep(time.Second)
+		err = DB.Ping()
+	}
+	if err != nil {
+		return err
+	}
+
+	files := strings.Split(dbScripts, ":")
+	err = RunSQL(files...)
+	if err != nil {
+		return err
+	}
+
+	slog.Info("Successfully connected!")
+	return nil
+}
+
+func RunSQL(files ...string) error {
+	for _, file := range files {
+		if file == "" {
+			continue
+		}
+		//https://stackoverflow.com/questions/12518876/how-to-check-if-a-file-exists-in-go
+		if _, err := os.Stat(file); err == nil {
+			fileBytes, err := os.ReadFile(file)
+			if err != nil {
+				return err
+			}
+
+			//https://stackoverflow.com/questions/12682405/strip-out-c-style-comments-from-a-byte
+			re := regexp.MustCompile("(?s)//.*?\n|/\\*.*?\\*/|(?s)--.*?\n|(?s)#.*?\n")
+			newBytes := re.ReplaceAll(fileBytes, nil)
+
+			requests := strings.Split(string(newBytes), ";")
+			for _, request := range requests {
+				request = strings.TrimSpace(request)
+				if len(request) > 0 {
+					_, err := DB.Exec(request)
+					if err != nil {
+						return fmt.Errorf("[%v] %v", request, err)
+					}
+				}
+			}
+		} else {
+			slog.Info("ignoring file %v (%v)", slog.String("file", file), slog.Any("error", err))
+		}
+	}
+	return nil
+}
+
+func CloseAndLog(c io.Closer) {
+	err := c.Close()
+	if err != nil {
+		slog.Info("could not close: %v", slog.Any("error", err))
 	}
 }
