@@ -2,41 +2,30 @@ package main
 
 import (
 	"bytes"
-	"context"
 	"crypto"
-	"crypto/rand"
 	"crypto/rsa"
 	"crypto/sha256"
 	"crypto/x509"
 	"encoding/base32"
 	"encoding/hex"
-	"encoding/json"
 	"flag"
 	"fmt"
+	"github.com/MatusOllah/slogcolor"
 	"github.com/dimiro1/banner"
-	dbLib "github.com/flatfeestack/go-lib/database"
-	env "github.com/flatfeestack/go-lib/environment"
-	prom "github.com/flatfeestack/go-lib/prometheus"
+	"github.com/fatih/color"
 	"github.com/go-jose/go-jose/v3"
 	"github.com/go-jose/go-jose/v3/jwt"
-	"github.com/gorilla/mux"
 	"github.com/joho/godotenv"
 	_ "github.com/lib/pq"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
-	log "github.com/sirupsen/logrus"
-	"github.com/xlzd/gotp"
 	"golang.org/x/crypto/ed25519"
 	"hash/fnv"
+	"log/slog"
 	rnd "math/rand"
 	"net"
 	"net/http"
-	"net/url"
 	"os"
-	"os/signal"
-	"regexp"
 	"strconv"
 	"strings"
-	"syscall"
 	"time"
 )
 
@@ -53,6 +42,15 @@ var (
 	secondsAdd   int
 	admins       []string
 	debug        bool
+	logger       = slog.New(slogcolor.NewHandler(os.Stderr, &slogcolor.Options{
+		Level:         slog.LevelDebug,
+		TimeFormat:    "15:04:05.000",
+		SrcFileMode:   slogcolor.ShortFile,
+		SrcFileLength: 16,
+		MsgPrefix:     color.HiWhiteString("|"),
+		MsgColor:      color.New(color.FgHiWhite),
+		MsgLength:     16,
+	}))
 )
 
 type Credentials struct {
@@ -123,8 +121,6 @@ type Opts struct {
 	EmailUrl        string
 	EmailToken      string
 	EmailLinkPrefix string
-	UrlSms          string
-	SmsToken        string
 	Audience        string
 	ExpireAccess    int
 	ExpireRefresh   int
@@ -148,8 +144,8 @@ type Opts struct {
 }
 
 func init() {
-	// Log as JSON instead of the default ASCII formatter.
-	log.SetFormatter(&log.JSONFormatter{})
+	color.NoColor = false
+	slog.SetDefault(logger)
 }
 
 func hash(s string) int64 {
@@ -160,62 +156,59 @@ func hash(s string) int64 {
 
 func NewOpts() *Opts {
 	opts := &Opts{}
-	flag.StringVar(&opts.Env, "env", env.LookupEnv("ENV"), "ENV variable")
-	flag.StringVar(&opts.Dev, "dev", env.LookupEnv("DEV"), "Dev settings with initial secret")
-	flag.StringVar(&opts.Issuer, "issuer", env.LookupEnv("ISSUER"), "name of issuer, default in dev is my-issuer")
-	flag.IntVar(&opts.Port, "port", env.LookupEnvInt("PORT",
+	flag.StringVar(&opts.Env, "env", LookupEnv("ENV"), "ENV variable")
+	flag.StringVar(&opts.Dev, "dev", LookupEnv("DEV"), "Dev settings with initial secret")
+	flag.StringVar(&opts.Issuer, "issuer", LookupEnv("ISSUER"), "name of issuer, default in dev is my-issuer")
+	flag.IntVar(&opts.Port, "port", LookupEnvInt("PORT",
 		8080), "listening HTTP port")
-	flag.StringVar(&opts.DBPath, "db-path", env.LookupEnv("DB_PATH",
+	flag.StringVar(&opts.DBPath, "db-path", LookupEnv("DB_PATH",
 		"./fastauth.db"), "DB path")
-	flag.StringVar(&opts.DBDriver, "db-driver", env.LookupEnv("DB_DRIVER",
+	flag.StringVar(&opts.DBDriver, "db-driver", LookupEnv("DB_DRIVER",
 		"sqlite3"), "DB driver")
-	flag.StringVar(&opts.DBScripts, "db-scripts", env.LookupEnv("DB_SCRIPTS"), "DB scripts to run at startup")
-	flag.StringVar(&opts.EmailFrom, "email-from", env.LookupEnv("EMAIL_FROM"), "Email from, default is info@flatfeestack.io")
-	flag.StringVar(&opts.EmailFromName, "email-from-name", env.LookupEnv("EMAIL_FROM_NAME",
+	flag.StringVar(&opts.DBScripts, "db-scripts", LookupEnv("DB_SCRIPTS"), "DB scripts to run at startup")
+	flag.StringVar(&opts.EmailFrom, "email-from", LookupEnv("EMAIL_FROM"), "Email from, default is info@flatfeestack.io")
+	flag.StringVar(&opts.EmailFromName, "email-from-name", LookupEnv("EMAIL_FROM_NAME",
 		"email@fastauth"), "Email from name, default is a empty string")
-	flag.StringVar(&opts.EmailUrl, "email-url", env.LookupEnv("EMAIL_URL"), "Email service URL")
-	flag.StringVar(&opts.EmailToken, "email-token", env.LookupEnv("EMAIL_TOKEN"), "Email service token")
-	flag.StringVar(&opts.EmailLinkPrefix, "email-prefix", env.LookupEnv("EMAIL_PREFIX"), "Email link prefix")
-	flag.StringVar(&opts.UrlSms, "sms-url", env.LookupEnv("SMS_URL"), "SMS service URL")
-	flag.StringVar(&opts.SmsToken, "sms-token", env.LookupEnv("SMS_TOKEN"), "SMS service token")
-	flag.StringVar(&opts.Audience, "audience", env.LookupEnv("AUDIENCE"), "Audience, default in dev is my-audience")
-	flag.IntVar(&opts.ExpireAccess, "expire-access", env.LookupEnvInt("EXPIRE_ACCESS",
+	flag.StringVar(&opts.EmailUrl, "email-url", LookupEnv("EMAIL_URL"), "Email service URL")
+	flag.StringVar(&opts.EmailToken, "email-token", LookupEnv("EMAIL_TOKEN"), "Email service token")
+	flag.StringVar(&opts.EmailLinkPrefix, "email-prefix", LookupEnv("EMAIL_PREFIX"), "Email link prefix")
+	flag.StringVar(&opts.Audience, "audience", LookupEnv("AUDIENCE"), "Audience, default in dev is my-audience")
+	flag.IntVar(&opts.ExpireAccess, "expire-access", LookupEnvInt("EXPIRE_ACCESS",
 		30*60), "Access token expiration in seconds, default 30min")
-	flag.IntVar(&opts.ExpireRefresh, "expire-refresh", env.LookupEnvInt("EXPIRE_REFRESH",
+	flag.IntVar(&opts.ExpireRefresh, "expire-refresh", LookupEnvInt("EXPIRE_REFRESH",
 		180*24*60*60), "Refresh token expiration in seconds, default 6month")
-	flag.IntVar(&opts.ExpireCode, "expire-code", env.LookupEnvInt("EXPIRE_CODE",
+	flag.IntVar(&opts.ExpireCode, "expire-code", LookupEnvInt("EXPIRE_CODE",
 		60), "Authtoken flow expiration in seconds, default 1min")
-	flag.StringVar(&opts.HS256, "hs256", env.LookupEnv("HS256"), "HS256 key")
-	flag.StringVar(&opts.RS256, "rs256", env.LookupEnv("RS256"), "RS256 key")
-	flag.StringVar(&opts.EdDSA, "eddsa", env.LookupEnv("EDDSA"), "EdDSA key")
-	flag.StringVar(&opts.OAuthUser, "oauth-user", env.LookupEnv("OAUTH_USER"), "Basic auth username for the server meta data")
-	flag.StringVar(&opts.OAuthPass, "oauth-pass", env.LookupEnv("OAUTH_PASS"), "Basic auth password for the server meta data")
-	flag.BoolVar(&opts.ResetRefresh, "reset-refresh", env.LookupEnv("RESET_REFRESH") != "", "Reset refresh token when setting the token")
-	flag.StringVar(&opts.Users, "users", env.LookupEnv("USERS"), "add these initial users. E.g, -users tom@test.ch:pw123;test@test.ch:123pw")
-	flag.BoolVar(&opts.AdminEndpoints, "admin-endpoints", env.LookupEnv("ADMIN_ENDPOINTS") != "", "Enable admin-facing endpoints. In dev mode these are enabled by default")
-	flag.BoolVar(&opts.UserEndpoints, "user-endpoints", env.LookupEnv("USER_ENDPOINTS") != "", "Enable user-facing endpoints. In dev mode these are enabled by default")
-	flag.BoolVar(&opts.OauthEndpoints, "oauth-enpoints", env.LookupEnv("OAUTH_ENDPOINTS") != "", "Enable oauth-facing endpoints. In dev mode these are enabled by default")
-	flag.BoolVar(&opts.DetailedError, "details", env.LookupEnv("DETAILS") != "", "Enable detailed errors")
-	flag.StringVar(&opts.Redirects, "redir", env.LookupEnv("REDIR"), "add client redirects. E.g, -redir clientId1:http://blabla;clientId2:http://blublu")
-	flag.BoolVar(&opts.PasswordFlow, "pwflow", env.LookupEnv("PWFLOW") != "", "enable password flow, default disabled")
-	flag.StringVar(&opts.Scope, "scope", env.LookupEnv("SCOPE"), "scope, default in dev is my-scope")
-	flag.StringVar(&opts.LogPath, "log", env.LookupEnv("LOG",
+	flag.StringVar(&opts.HS256, "hs256", LookupEnv("HS256"), "HS256 key")
+	flag.StringVar(&opts.RS256, "rs256", LookupEnv("RS256"), "RS256 key")
+	flag.StringVar(&opts.EdDSA, "eddsa", LookupEnv("EDDSA"), "EdDSA key")
+	flag.StringVar(&opts.OAuthUser, "oauth-user", LookupEnv("OAUTH_USER"), "Basic auth username for the server meta data")
+	flag.StringVar(&opts.OAuthPass, "oauth-pass", LookupEnv("OAUTH_PASS"), "Basic auth password for the server meta data")
+	flag.BoolVar(&opts.ResetRefresh, "reset-refresh", LookupEnv("RESET_REFRESH") != "", "Reset refresh token when setting the token")
+	flag.StringVar(&opts.Users, "users", LookupEnv("USERS"), "add these initial users. E.g, -users tom@test.ch:pw123;test@test.ch:123pw")
+	flag.BoolVar(&opts.AdminEndpoints, "admin-endpoints", LookupEnv("ADMIN_ENDPOINTS") != "", "Enable admin-facing endpoints. In dev mode these are enabled by default")
+	flag.BoolVar(&opts.UserEndpoints, "user-endpoints", LookupEnv("USER_ENDPOINTS") != "", "Enable user-facing endpoints. In dev mode these are enabled by default")
+	flag.BoolVar(&opts.OauthEndpoints, "oauth-enpoints", LookupEnv("OAUTH_ENDPOINTS") != "", "Enable oauth-facing endpoints. In dev mode these are enabled by default")
+	flag.BoolVar(&opts.DetailedError, "details", LookupEnv("DETAILS") != "", "Enable detailed errors")
+	flag.StringVar(&opts.Redirects, "redir", LookupEnv("REDIR"), "add client redirects. E.g, -redir clientId1:http://blabla;clientId2:http://blublu")
+	flag.BoolVar(&opts.PasswordFlow, "pwflow", LookupEnv("PWFLOW") != "", "enable password flow, default disabled")
+	flag.StringVar(&opts.Scope, "scope", LookupEnv("SCOPE"), "scope, default in dev is my-scope")
+	flag.StringVar(&opts.LogPath, "log", LookupEnv("LOG",
 		os.TempDir()+"/ffs"), "Log directory, default is /tmp/ffs")
-	flag.StringVar(&opts.Admins, "admins", env.LookupEnv("ADMINS"), "Admins")
+	flag.StringVar(&opts.Admins, "admins", LookupEnv("ADMINS"), "Admins")
 
 	flag.Usage = func() {
-		fmt.Fprintf(flag.CommandLine.Output(), "Usage of %s:\n", os.Args[0])
+		fmt.Printf("Usage of %s:\n", os.Args[0])
 		flag.PrintDefaults()
 	}
-
 	flag.Parse()
 
 	//set defaults, be explicit
 	if opts.Env == "local" || opts.Env == "dev" {
 		debug = true
-		log.SetLevel(log.DebugLevel)
+		slog.SetLogLoggerLevel(slog.LevelDebug)
 	} else {
-		log.SetLevel(log.InfoLevel)
+		slog.SetLogLoggerLevel(slog.LevelInfo)
 	}
 
 	//set defaults
@@ -232,25 +225,22 @@ func NewOpts() *Opts {
 		if opts.EmailUrl == "" {
 			opts.EmailUrl = "http://localhost:8080/send/email/{email}/{token}"
 		}
-		if opts.UrlSms == "" {
-			opts.UrlSms = "http://localhost:8080/send/sms/{sms}/{token}"
-		}
 
-		if strings.ToLower(opts.HS256) != "true" && strings.ToLower(opts.EdDSA) != "true" {
+		if strings.ToLower(opts.RS256) == "true" {
 			//work around this issue: https://github.com/golang/go/issues/38548
 			//we want for testing to have the same key, I don't care for any database keys
 			rsaPrivKey1, err := rsa.GenerateKey(rnd.New(rnd.NewSource(hash(opts.Dev))), 2048)
 			if err != nil {
-				log.Fatalf("cannot generate rsa key %v", err)
+				slog.Error("cannot generate rsa key", slog.Any("error", err))
 			}
 			rsaPrivKey2, err := rsa.GenerateKey(rnd.New(rnd.NewSource(hash(opts.Dev))), 2048)
 			if err != nil {
-				log.Fatalf("cannot generate rsa key %v", err)
+				slog.Error("cannot generate rsa key", slog.Any("error", err))
 			}
 			for rsaPrivKey1.Equal(rsaPrivKey2) {
 				rsaPrivKey2, err = rsa.GenerateKey(rnd.New(rnd.NewSource(hash(opts.Dev))), 2048)
 				if err != nil {
-					log.Fatalf("cannot generate rsa key %v", err)
+					slog.Error("cannot generate rsa key", slog.Any("error", err))
 				}
 			}
 			rsaPrivKey := rsaPrivKey1
@@ -259,19 +249,26 @@ func NewOpts() *Opts {
 			}
 
 			if err != nil {
-				log.Fatalf("cannot generate rsa key %v", err)
+				slog.Error("cannot generate rsa key", slog.Any("error", err))
 			}
 			encPrivRSA, err := x509.MarshalPKCS8PrivateKey(rsaPrivKey)
 			if err != nil {
-				log.Fatalf("cannot generate rsa key %v", err)
+				slog.Error("cannot generate rsa key", slog.Any("error", err))
 			}
 			opts.RS256 = base32.StdEncoding.EncodeToString(encPrivRSA)
-		} else if strings.ToLower(opts.HS256) != "true" && strings.ToLower(opts.RS256) != "true" {
+		} else if strings.ToLower(opts.EdDSA) == "true" {
 			_, edPrivKey, err := ed25519.GenerateKey(rnd.New(rnd.NewSource(hash(opts.Dev))))
 			if err != nil {
-				log.Fatalf("cannot generate eddsa key %v", err)
+				slog.Error("cannot generate eddsa key", slog.Any("error", err))
 			}
 			opts.EdDSA = base32.StdEncoding.EncodeToString(edPrivKey)
+		} else if strings.ToLower(opts.HS256) != "true" && opts.HS256 != "" {
+			slog.Warn("DEV mode enabled, ignoring key", slog.String("opts.HS256", opts.HS256))
+		} else {
+			h := sha256.New()
+			h.Write([]byte(opts.Dev))
+			jwtKey = h.Sum(nil)
+			opts.HS256 = base32.StdEncoding.EncodeToString(jwtKey)
 		}
 		if opts.OAuthUser == "" {
 			opts.OAuthUser = "clientId"
@@ -289,9 +286,9 @@ func NewOpts() *Opts {
 			opts.Users = "user:pass"
 		}
 
-		log.Printf("DEV mode active, key is %v, hex(%v)", opts.Dev, opts.HS256)
-		log.Printf("DEV mode active, rsa is hex(%v)", opts.RS256)
-		log.Printf("DEV mode active, eddsa is hex(%v)", opts.EdDSA)
+		slog.Debug("DEV mode active, key is %v, hex(%v)", slog.String("opts.Dev", opts.Dev), slog.String("opts.HS256", opts.HS256))
+		slog.Debug("DEV mode active, rsa is hex(%v)", slog.String("opts.RS256", opts.RS256))
+		slog.Debug("DEV mode active, eddsa is hex(%v)", slog.String("opts.EdDSA", opts.EdDSA))
 	}
 
 	admins = strings.Split(opts.Admins, ";")
@@ -299,57 +296,70 @@ func NewOpts() *Opts {
 		admins = append(admins, opts.OAuthUser)
 	}
 
-	if opts.HS256 == "" && opts.RS256 == "" && opts.EdDSA == "" {
-		fmt.Printf("Paramter hs256, rs256, or eddsa not set. One of them is mandatory. Choose one\n")
+	// Check that exactly one of HS256, RS256, or EdDSA is set.
+	count := 0
+	if opts.HS256 != "" {
+		count += 1
+	}
+
+	if opts.RS256 != "" {
+		count += 1
+	}
+
+	if opts.EdDSA != "" {
+		count += 1
+	}
+	if count != 1 {
+		fmt.Printf("Exactly one of hs256, rs256, or eddsa must be set. Choose one\n")
 		flag.PrintDefaults()
 		os.Exit(1)
 	}
 
-	if opts.HS256 == "true" {
-		opts.HS256 = base32.StdEncoding.EncodeToString([]byte(opts.Dev))
-	} else if opts.HS256 != "" {
+	if opts.HS256 != "" {
 		var err error
 		jwtKey, err = base32.StdEncoding.DecodeString(opts.HS256)
 		if err != nil {
-			h := sha256.New()
-			h.Write([]byte(opts.HS256))
-			jwtKey = h.Sum(nil)
+			slog.Error("cannot hash", slog.Any("error", err))
 		}
 	}
 
 	if opts.RS256 != "" {
 		rsaDec, err := base32.StdEncoding.DecodeString(opts.RS256)
 		if err != nil {
-			log.Fatalf("cannot decode %v", opts.RS256)
+			slog.Error("cannot decode", slog.String("opts.RS256", opts.RS256))
 		}
 		i, err := x509.ParsePKCS8PrivateKey(rsaDec)
 		if err != nil {
-			log.Fatalf("cannot decode %v", rsaDec)
+			slog.Error("cannot create private key", slog.String("rsaDec", opts.RS256))
 		}
 		privRSA = i.(*rsa.PrivateKey)
 		k := jose.JSONWebKey{Key: privRSA.Public()}
 		kid, err := k.Thumbprint(crypto.SHA256)
 		if err != nil {
-			log.Fatalf("cannot thumb rsa %v", err)
+			slog.Error("cannot thumb rsa", slog.Any("error", err))
 		}
 		privRSAKid = hex.EncodeToString(kid)
-		log.Infof("kid: %v", privRSAKid)
+		slog.Info("kid", slog.String("privRSAKid", privRSAKid))
 	}
 
 	if opts.EdDSA != "" {
 		eddsa, err := base32.StdEncoding.DecodeString(opts.EdDSA)
 		if err != nil {
-			log.Fatalf("cannot decode %v", opts.EdDSA)
+			slog.Error("cannot decode", slog.String("opts.EdDSA", opts.EdDSA))
 		}
 		privEdDSA0 := ed25519.PrivateKey(eddsa)
 		privEdDSA = &privEdDSA0
 		k := jose.JSONWebKey{Key: privEdDSA.Public()}
 		kid, err := k.Thumbprint(crypto.SHA256)
 		if err != nil {
-			log.Fatalf("cannot thumb eddsa %v", err)
+			slog.Error("cannot thumb eddsa", slog.Any("error", err))
 		}
 		privEdDSAKid = hex.EncodeToString(kid)
 	}
+
+	tokenExp = time.Second * time.Duration(opts.ExpireAccess)
+	refreshExp = time.Second * time.Duration(opts.ExpireRefresh)
+	codeExp = time.Second * time.Duration(opts.ExpireCode)
 
 	return opts
 }
@@ -387,85 +397,105 @@ func checkEmailPassword(email string, password string) (*dbRes, string, error) {
 	return result, "", nil
 }
 
-func serverRest(keepAlive bool) (*http.Server, <-chan bool, error) {
-	tokenExp = time.Second * time.Duration(opts.ExpireAccess)
-	refreshExp = time.Second * time.Duration(opts.ExpireRefresh)
-	codeExp = time.Second * time.Duration(opts.ExpireCode)
+func middlewareLog(handlerFunc func(http.ResponseWriter, *http.Request)) func(w http.ResponseWriter, r *http.Request) {
+	loggedHandler := logRequestHandler(http.HandlerFunc(handlerFunc))
+	return loggedHandler
+}
 
-	router := mux.NewRouter()
-	router.Use(func(next http.Handler) http.Handler {
-		return logRequestHandler(next)
-	})
-	router.Use(prom.PrometheusMiddleware)
-	registry := prom.CreateRegistry()
-	router.Path("/metrics").Handler(promhttp.HandlerFor(
-		registry,
-		promhttp.HandlerOpts{
-			Registry: registry,
-			// Opt into OpenMetrics to support exemplars.
-			EnableOpenMetrics: true,
-		},
-	))
+func middlewareJwtLog(handlerFunc func(http.ResponseWriter, *http.Request, *TokenClaims)) func(w http.ResponseWriter, r *http.Request) {
+	jh := jwtAuth(handlerFunc)
+	return middlewareLog(jh)
+}
+
+func middlewareJwtAdminLog(handlerFunc func(http.ResponseWriter, *http.Request, string)) func(w http.ResponseWriter, r *http.Request) {
+	jh := jwtAuthAdmin(handlerFunc, admins)
+	return middlewareLog(jh)
+}
+
+func setupRouter() *http.ServeMux {
+	router := http.NewServeMux()
 
 	if opts.UserEndpoints {
-		router.HandleFunc("/login", login).Methods(http.MethodPost)
-		router.HandleFunc("/refresh", refresh).Methods(http.MethodPost)
-		router.HandleFunc("/signup", signup).Methods(http.MethodPost)
-		router.HandleFunc("/reset/{email}", resetEmail).Methods(http.MethodPost)
-		router.HandleFunc("/confirm/signup/{email}/{token}", confirmEmail).Methods(http.MethodGet)
-		router.HandleFunc("/confirm/signup", confirmEmailPost).Methods(http.MethodPost)
-		router.HandleFunc("/confirm/invite", confirmInvite).Methods(http.MethodPost)
-		router.HandleFunc("/confirm/reset", confirmReset).Methods(http.MethodPost)
+		router.HandleFunc("POST /login", middlewareLog(login))
+		router.HandleFunc("POST /refresh", middlewareLog(refresh))
+		router.HandleFunc("POST /signup", middlewareLog(signup))
+		router.HandleFunc("POST /reset/{email}", middlewareLog(resetEmail))
+		router.HandleFunc("GET /confirm/signup/{email}/{token}", middlewareLog(confirmEmail))
+		router.HandleFunc("POST /confirm/signup", middlewareLog(confirmEmailPost))
+		router.HandleFunc("POST /confirm/invite", middlewareLog(confirmInvite))
+		router.HandleFunc("POST /confirm/reset", middlewareLog(confirmReset))
 
-		router.HandleFunc("/invite/{email}", jwtAuth(invite)).Methods(http.MethodPost)
-		router.HandleFunc("/setup/totp", jwtAuth(setupTOTP)).Methods(http.MethodPost)
-		router.HandleFunc("/confirm/totp/{token}", jwtAuth(confirmTOTP)).Methods(http.MethodPost)
-		router.HandleFunc("/setup/sms/{sms}", jwtAuth(setupSMS)).Methods(http.MethodPost)
-		router.HandleFunc("/confirm/sms/{token}", jwtAuth(confirmSMS)).Methods(http.MethodPost)
+		router.HandleFunc("POST /invite/{email}", middlewareJwtLog(invite))
+		router.HandleFunc("POST /setup/totp", middlewareJwtLog(setupTOTP))
+		router.HandleFunc("POST /confirm/totp/{token}", middlewareJwtLog(confirmTOTP))
 	}
 	//logout
-	router.HandleFunc("/authen/logout", jwtAuth(logout)).Methods(http.MethodGet)
+	router.HandleFunc("GET /authen/logout", middlewareJwtLog(logout))
 
 	//maintenance stuff
-	router.HandleFunc("/readiness", readiness).Methods(http.MethodGet)
-	router.HandleFunc("/liveness", liveness).Methods(http.MethodGet)
+	router.HandleFunc("GET /readiness", middlewareLog(readiness))
+	router.HandleFunc("GET /liveness", middlewareLog(liveness))
 
 	//display for debug and testing
 	if debug {
-		router.HandleFunc("/send/email/{email}/{token}", displayEmail).Methods(http.MethodPost)
-		router.HandleFunc("/send/sms/{sms}/{token}", displaySMS).Methods(http.MethodPost)
 		//admin endpoints
-		router.HandleFunc("/admin/time", jwtAuthAdmin(serverTime, admins)).Methods(http.MethodGet)
-		router.HandleFunc("/admin/timewarp/{hours}", jwtAuthAdmin(timeWarp, admins)).Methods(http.MethodPost)
+		router.HandleFunc("GET /admin/time", middlewareJwtAdminLog(serverTime))
+		router.HandleFunc("POST /admin/timewarp/{hours}", middlewareJwtAdminLog(timeWarp))
 	}
 
 	if opts.OauthEndpoints {
-		router.HandleFunc("/oauth/login", login).Methods(http.MethodPost)
-		//TODO: check regitier
-		router.HandleFunc("/ap/login", login).Methods(http.MethodPost)
-		router.HandleFunc("/oauth/token", oauth).Methods(http.MethodPost)
-		router.HandleFunc("/oauth/revoke", revoke).Methods(http.MethodPost)
-		router.HandleFunc("/oauth/authorize", authorize).Methods(http.MethodGet, http.MethodOptions)
+		router.HandleFunc("POST /oauth/login", middlewareLog(login))
+		router.HandleFunc("POST /oauth/token", middlewareLog(oauth))
+		router.HandleFunc("POST /oauth/revoke", middlewareLog(revoke))
+		router.HandleFunc("GET /oauth/authorize", middlewareLog(authorize))
 		//convenience function
 		if opts.Env == "dev" || opts.Env == "local" {
-			router.HandleFunc("/", authorize).Methods(http.MethodGet)
-			//TODO: check regitier
-			router.HandleFunc("/ap/registration", authorize).Methods(http.MethodGet)
+			router.HandleFunc("GET /", middlewareLog(authorize))
 		}
-		//router.HandleFunc("/oauth/.well-known/jwks.json", jwkFunc).Methods(http.MethodGet)
-		router.HandleFunc("/.well-known/jwks.json", jwkFunc).Methods(http.MethodGet)
+		router.HandleFunc("GET /.well-known/jwks.json", middlewareLog(jwkFunc))
 	}
 
 	if opts.AdminEndpoints {
-		router.HandleFunc("/users/usernames/{email}/cancellation", jwtAuthAdmin(deleteUser, admins)).Methods(http.MethodPost)
-		router.HandleFunc("/users/usernames/{email}/attributes", jwtAuthAdmin(updateUser, admins)).Methods(http.MethodPatch)
-		router.HandleFunc("/admin/login-as/{email}", jwtAuthAdmin(asUser, admins)).Methods(http.MethodPost)
+		router.HandleFunc("POST /users/usernames/{email}/cancellation", middlewareJwtAdminLog(deleteUser))
+		router.HandleFunc("PATCH /users/usernames/{email}/attributes", middlewareJwtAdminLog(updateUser))
+		router.HandleFunc("POST /admin/login-as/{email}", middlewareJwtAdminLog(asUser))
 	}
 
-	router.NotFoundHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		log.Printf("[404] no route matched for: %s, %s", r.URL, r.Method)
+	router.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
+		slog.Info("[404] no route matched for", "url", r.URL, "method", r.Method)
 		w.WriteHeader(http.StatusNotFound)
 	})
+
+	return router
+}
+
+func main() {
+	//the .env should be loaded before showing the banner, as the banner shows also the ENVs
+	err := godotenv.Load()
+	if err != nil {
+		slog.Info("Could not find .env file, using defaults")
+	}
+
+	opts = NewOpts()
+
+	f, err := os.Open("banner.txt")
+	if err == nil {
+		banner.Init(os.Stdout, true, false, f)
+	} else {
+		slog.Info("could not display banner...")
+	}
+
+	err = InitDb(opts.DBDriver, opts.DBPath, opts.DBScripts)
+	if err != nil {
+		slog.Error("DB not initialized",
+			slog.Any("error", err))
+		os.Exit(1)
+	}
+	defer CloseDb()
+
+	addInitialUsers()
+
+	router := setupRouter()
 
 	s := &http.Server{
 		Addr:         ":" + strconv.Itoa(opts.Port),
@@ -473,210 +503,15 @@ func serverRest(keepAlive bool) (*http.Server, <-chan bool, error) {
 		ReadTimeout:  5 * time.Second,
 		WriteTimeout: 10 * time.Second,
 	}
-	s.SetKeepAlivesEnabled(keepAlive)
 
 	l, err := net.Listen("tcp", s.Addr)
 	if err != nil {
-		return nil, nil, err
+		slog.Error("Listen failed",
+			slog.Any("error", err))
 	}
 
-	done := make(chan bool)
-	go func(s *http.Server, l net.Listener) {
-		log.Printf("Starting auth server on port %v...", s.Addr)
-		if err := s.Serve(l); err != nil && err != http.ErrServerClosed {
-			log.Fatal(err)
-		}
-		log.Printf("Shutdown\n")
-		done <- true
-	}(s, l)
-	return s, done, nil
-}
-
-////////// Util functions
-
-func genRnd(n int) ([]byte, error) {
-	b := make([]byte, n)
-	_, err := rand.Read(b)
-	// Note that err == nil only if we read len(b) bytes.
-	if err != nil {
-		return nil, err
+	if err := s.Serve(l); err != nil && err != http.ErrServerClosed {
+		slog.Error("Server stopped",
+			slog.Any("error", err))
 	}
-
-	return b, nil
-}
-
-func genToken() (string, error) {
-	rnd, err := genRnd(20)
-	if err != nil {
-		return "", err
-	}
-	return base32.StdEncoding.EncodeToString(rnd), nil
-}
-
-func WriteErrorf(w http.ResponseWriter, code int, format string, a ...interface{}) {
-	msg := fmt.Sprintf(format, a...)
-	log.Errorf(msg)
-	w.Header().Set("Content-Type", "application/json;charset=UTF-8")
-	w.Header().Set("Cache-Control", "no-store")
-	w.Header().Set("Pragma", "no-cache")
-	w.WriteHeader(code)
-
-	msgEnc, err := json.Marshal(msg)
-	if err != nil {
-		log.Errorf("error while trying to encode msg:  %v, err: %v", msg, err)
-		return
-	}
-	_, err = w.Write([]byte(`{"error":` + string(msgEnc) + `}`))
-	if err != nil {
-		log.Errorf("Something went wrong while writing error message: %v", err)
-		return
-	}
-}
-
-func sendSMS(url string) error {
-	c := &http.Client{
-		Timeout: 15 * time.Second,
-	}
-	resp, err := c.Get(url)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("could not update DB as status from email server: %v %v", resp.Status, resp.StatusCode)
-	}
-	return nil
-}
-
-func validateEmail(email string) error {
-	var rxEmail = regexp.MustCompile("^[a-zA-Z0-9.!#$%&'*+\\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$")
-
-	if len(email) > 254 || !rxEmail.MatchString(email) {
-		return fmt.Errorf("[%s] is not a valid email address", email)
-	}
-	return nil
-}
-
-func validatePassword(password string) error {
-	if len(password) < 8 {
-		return fmt.Errorf("password is less than 8 characters")
-	}
-	return nil
-}
-
-func newTOTP(secret string) *gotp.TOTP {
-	hasher := &gotp.Hasher{
-		HashName: "sha256",
-		Digest:   sha256.New,
-	}
-	return gotp.NewTOTP(secret, 6, 30, hasher)
-}
-
-func basicAuth(r *http.Request) (string, error) {
-	if opts.OAuthUser != "" || opts.OAuthPass != "" {
-		user, pass, ok := r.BasicAuth()
-		if !ok || user != opts.OAuthUser || pass != opts.OAuthPass {
-			clientId, err := param("client_id", r)
-			if err != nil {
-				return "", fmt.Errorf("no client_id %v", err)
-			}
-			clientSecret, err := param("client_secret", r)
-			if err != nil {
-				return "", fmt.Errorf("no client_secret %v", err)
-			}
-			if clientId != opts.OAuthUser || clientSecret != opts.OAuthPass {
-				return "", fmt.Errorf("no match, user/pass %v", err)
-			}
-		}
-		return user, nil
-	}
-	return "", fmt.Errorf("no user/pass set")
-}
-
-func param(name string, r *http.Request) (string, error) {
-	n1 := mux.Vars(r)[name]
-	n2, err := url.QueryUnescape(r.URL.Query().Get(name))
-	if err != nil {
-		return "", err
-	}
-	err = r.ParseForm()
-	if err != nil {
-		return "", err
-	}
-	n3 := r.FormValue(name)
-
-	if n1 == "" {
-		if n2 == "" {
-			return n3, nil
-		}
-		return n2, nil
-	}
-	return n1, nil
-}
-
-func paramJson(name string, r *http.Request) (string, error) {
-	var objmap map[string]json.RawMessage
-	err := json.NewDecoder(r.Body).Decode(&objmap)
-	if err != nil {
-		return "", err
-	}
-	var s string
-	err = json.Unmarshal(objmap[name], &s)
-	if err != nil {
-		return "", err
-	}
-	return s, nil
-}
-
-func timeNow() time.Time {
-	if debug {
-		return time.Now().Add(time.Duration(secondsAdd) * time.Second).UTC()
-	} else {
-		return time.Now().UTC()
-	}
-}
-
-func main() {
-	//the .env should be loaded before showing the banner, as the banner shows also the ENVs
-	err := godotenv.Load()
-	if err != nil {
-		log.Printf("Could not find env file [%v], using defaults", err)
-	}
-
-	f, err := os.Open("banner.txt")
-	if err == nil {
-		banner.Init(os.Stdout, true, false, f)
-	} else {
-		log.Printf("could not display banner...")
-	}
-
-	opts = NewOpts()
-
-	//logs: we have to ensure the directory we want to write to
-	// already exists
-	err = os.MkdirAll(opts.LogPath, 0755)
-	if err != nil {
-		log.Fatalf("os.MkdirAll()")
-	}
-
-	//db: init the database
-	err = dbLib.InitDb(opts.DBDriver, opts.DBPath, opts.DBScripts)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer dbLib.DB.Close()
-	addInitialUsers()
-	serverRest, doneChannelRest, err := serverRest(true)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	c := make(chan os.Signal)
-	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
-	go func() {
-		<-c
-		serverRest.Shutdown(context.Background())
-	}()
-
-	<-doneChannelRest
 }
