@@ -4,7 +4,6 @@ import (
 	"database/sql"
 	"encoding/json"
 	"fmt"
-	dbLib "github.com/flatfeestack/go-lib/database"
 	"github.com/google/uuid"
 	"time"
 )
@@ -30,11 +29,11 @@ type AnalysisResponse struct {
 }
 
 func InsertAnalysisRequest(a AnalysisRequest, now time.Time) error {
-	stmt, err := dbLib.DB.Prepare("INSERT INTO analysis_request(id, repo_id, date_from, date_to, git_url, created_at) VALUES($1, $2, $3, $4, $5, $6)")
+	stmt, err := DB.Prepare("INSERT INTO analysis_request(id, repo_id, date_from, date_to, git_url, created_at) VALUES($1, $2, $3, $4, $5, $6)")
 	if err != nil {
 		return fmt.Errorf("prepare INSERT INTO analysis_request for %v statement event: %v", a.Id, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	var res sql.Result
 	res, err = stmt.Exec(a.Id, a.RepoId, a.DateFrom, a.DateTo, a.GitUrl, now.Format("2006-01-02"))
@@ -51,13 +50,13 @@ func InsertAnalysisResponse(reqId uuid.UUID, gitEmail string, names []string, we
 		return fmt.Errorf("cannot marshal %v", err)
 	}
 
-	stmt, err := dbLib.DB.Prepare(`INSERT INTO analysis_response(
+	stmt, err := DB.Prepare(`INSERT INTO analysis_response(
                                      id, analysis_request_id, git_email, git_names, weight, created_at) 
 									 VALUES ($1, $2, $3, $4, $5, $6)`)
 	if err != nil {
 		return fmt.Errorf("prepare INSERT INTO analysis_response for %v statement event: %v", reqId, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	var res sql.Result
 	res, err = stmt.Exec(uuid.New(), reqId, gitEmail, string(bytes), weight, now)
@@ -73,7 +72,7 @@ func FindLatestAnalysisRequest(repoId uuid.UUID) (*AnalysisRequest, error) {
 	var a AnalysisRequest
 
 	//https://stackoverflow.com/questions/47479973/golang-postgresql-array#47480256
-	err := dbLib.DB.
+	err := DB.
 		QueryRow(`SELECT id, repo_id, date_from, date_to, git_url, received_at, error FROM (
                           SELECT id, repo_id, date_from, date_to, git_url, received_at, error,
                             RANK() OVER (PARTITION BY repo_id ORDER BY date_to DESC) dest_rank
@@ -94,7 +93,7 @@ func FindLatestAnalysisRequest(repoId uuid.UUID) (*AnalysisRequest, error) {
 func FindAllLatestAnalysisRequest(dateTo time.Time) ([]AnalysisRequest, error) {
 	var as []AnalysisRequest
 
-	rows, err := dbLib.DB.Query(`SELECT id, repo_id, date_from, date_to, git_url, received_at, error FROM (
+	rows, err := DB.Query(`SELECT id, repo_id, date_from, date_to, git_url, received_at, error FROM (
                           SELECT id, repo_id, date_from, date_to, git_url, received_at, error,
                             RANK() OVER (PARTITION BY repo_id ORDER BY date_to DESC) dest_rank
                             FROM analysis_request) AS x
@@ -104,7 +103,7 @@ func FindAllLatestAnalysisRequest(dateTo time.Time) ([]AnalysisRequest, error) {
 	if err != nil {
 		return nil, err
 	}
-	defer dbLib.CloseAndLog(rows)
+	defer CloseAndLog(rows)
 
 	for rows.Next() {
 		var a AnalysisRequest
@@ -118,12 +117,11 @@ func FindAllLatestAnalysisRequest(dateTo time.Time) ([]AnalysisRequest, error) {
 }
 
 func UpdateAnalysisRequest(reqId uuid.UUID, now time.Time, errStr *string) error {
-	//stmt, err := dbLib.DB.Prepare(`UPDATE analysis_request set received_at = $2, error = $3 WHERE id = $1`)
-	stmt, err := dbLib.DB.Prepare(`UPDATE analysis_request set received_at = $1, error = $2 WHERE id = $3`)
+	stmt, err := DB.Prepare(`UPDATE analysis_request set received_at = $1, error = $2 WHERE id = $3`)
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE analysis_request for statement event: %v", err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	var res sql.Result
 	res, err = stmt.Exec(now, errStr, reqId)
@@ -136,14 +134,14 @@ func UpdateAnalysisRequest(reqId uuid.UUID, now time.Time, errStr *string) error
 func FindAnalysisResults(reqId uuid.UUID) ([]AnalysisResponse, error) {
 	var ars []AnalysisResponse
 
-	rows, err := dbLib.DB.Query(`SELECT id, git_email, git_names, weight
+	rows, err := DB.Query(`SELECT id, git_email, git_names, weight
                                  FROM analysis_response 
                                  WHERE analysis_request_id = $1`, reqId)
 
 	if err != nil {
 		return nil, err
 	}
-	defer dbLib.CloseAndLog(rows)
+	defer CloseAndLog(rows)
 
 	for rows.Next() {
 		var ar AnalysisResponse
@@ -166,7 +164,7 @@ func FindAnalysisResults(reqId uuid.UUID) ([]AnalysisResponse, error) {
 func FindRepoContribution(repoId uuid.UUID) ([]Contributions, error) {
 	var cs []Contributions
 
-	rows, err := dbLib.DB.Query(`SELECT areq.date_from, areq.date_to, ares.git_email, ares.git_names, ares.weight
+	rows, err := DB.Query(`SELECT areq.date_from, areq.date_to, ares.git_email, ares.git_names, ares.weight
                         FROM analysis_request areq
                         INNER JOIN analysis_response ares on areq.id = ares.analysis_request_id
                         WHERE areq.repo_id=$1 AND areq.error IS NULL 
@@ -175,7 +173,7 @@ func FindRepoContribution(repoId uuid.UUID) ([]Contributions, error) {
 	if err != nil {
 		return nil, err
 	}
-	defer dbLib.CloseAndLog(rows)
+	defer CloseAndLog(rows)
 
 	for rows.Next() {
 		var c Contributions
@@ -198,7 +196,7 @@ func FindRepoContribution(repoId uuid.UUID) ([]Contributions, error) {
 
 func FindRepoContributors(repoId uuid.UUID) (int, error) {
 	var c int
-	err := dbLib.DB.QueryRow(`SELECT count(distinct git_email) as c
+	err := DB.QueryRow(`SELECT count(distinct git_email) as c
                         FROM analysis_request areq
                         INNER JOIN analysis_response ares on areq.id = ares.analysis_request_id
                         WHERE areq.repo_id=$1 AND areq.error IS NULL`, repoId).Scan(&c)
