@@ -1,7 +1,6 @@
 package db
 
 import (
-	"backend/pkg/util"
 	"fmt"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
@@ -10,8 +9,8 @@ import (
 )
 
 func TestAnalysisRequest(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	r := Repo{
 		Id:          uuid.New(),
@@ -45,8 +44,8 @@ func TestAnalysisRequest(t *testing.T) {
 }
 
 func TestAnalysisRequest2(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	r1 := Repo{
 		Id:          uuid.New(),
@@ -150,8 +149,8 @@ func TestAnalysisRequest2(t *testing.T) {
 }
 
 func TestAnalysisResponse(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	r := Repo{
 		Id:          uuid.New(),
@@ -197,8 +196,8 @@ func TestAnalysisResponse(t *testing.T) {
 }
 
 func TestFindRepoContribution(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	r := insertTestRepo(t)
 
 	a := AnalysisRequest{
