@@ -1,7 +1,6 @@
 package db
 
 import (
-	"backend/pkg/util"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"math/big"
@@ -10,8 +9,8 @@ import (
 )
 
 func TestGitEmailInsert(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	uid := insertTestUser(t, "email1").Id
 
@@ -24,8 +23,8 @@ func TestGitEmailInsert(t *testing.T) {
 }
 
 func TestCountExistingOrConfirmedGitEmail(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	uid := insertTestUser(t, "foo@bar.ch").Id
 	uid2 := insertTestUser(t, "bar@foo.ch").Id
@@ -52,8 +51,8 @@ func TestCountExistingOrConfirmedGitEmail(t *testing.T) {
 }
 
 func TestGitEmailFind(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	uid := insertTestUser(t, "email1").Id
 
@@ -67,8 +66,8 @@ func TestGitEmailFind(t *testing.T) {
 }
 
 func TestGitEmailDelete(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	uid := insertTestUser(t, "email1").Id
 
@@ -87,8 +86,8 @@ func TestGitEmailDelete(t *testing.T) {
 }
 
 func TestGitEmailFromUserEmailsSentDelete(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	uid := insertTestUser(t, "foo@bar.ch").Id
 
@@ -105,8 +104,8 @@ func TestGitEmailFromUserEmailsSentDelete(t *testing.T) {
 }
 
 func TestGitEmailConfirm(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	uid := insertTestUser(t, "email1").Id
 
@@ -120,8 +119,8 @@ func TestGitEmailConfirm(t *testing.T) {
 }
 
 func TestInsertUnclaimed(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 
 	r := insertTestRepo(t)
 
