@@ -3,7 +3,6 @@ package db
 import (
 	"database/sql"
 	"fmt"
-	dbLib "github.com/flatfeestack/go-lib/database"
 	"github.com/google/uuid"
 	"time"
 )
@@ -20,13 +19,13 @@ func FindInvitationsByAnyEmail(email string) ([]Invite, error) {
 	query := `SELECT from_email, to_email, confirmed_at, created_at 
               FROM invite 
               WHERE to_email=$1 OR from_email=$1`
-	rows, err := dbLib.DB.Query(query, email)
+	rows, err := DB.Query(query, email)
 
 	switch err {
 	case sql.ErrNoRows:
 		return nil, nil
 	case nil:
-		defer dbLib.CloseAndLog(rows)
+		defer CloseAndLog(rows)
 		for rows.Next() {
 			var inv Invite
 			err = rows.Scan(&inv.Email, &inv.InviteEmail, &inv.ConfirmedAt, &inv.CreatedAt)
@@ -46,13 +45,13 @@ func FindMyInvitations(email string) ([]Invite, error) {
 	query := `SELECT from_email, to_email, confirmed_at, created_at 
               FROM invite 
               WHERE from_email=$1`
-	rows, err := dbLib.DB.Query(query, email)
+	rows, err := DB.Query(query, email)
 
 	switch err {
 	case sql.ErrNoRows:
 		return nil, nil
 	case nil:
-		defer dbLib.CloseAndLog(rows)
+		defer CloseAndLog(rows)
 		for rows.Next() {
 			var inv Invite
 			err = rows.Scan(&inv.Email, &inv.InviteEmail, &inv.ConfirmedAt, &inv.CreatedAt)
@@ -68,11 +67,11 @@ func FindMyInvitations(email string) ([]Invite, error) {
 }
 
 func DeleteInvite(fromEmail string, toEmail string) error {
-	stmt, err := dbLib.DB.Prepare("DELETE FROM invite WHERE from_email = $1 AND to_email = $2")
+	stmt, err := DB.Prepare("DELETE FROM invite WHERE from_email = $1 AND to_email = $2")
 	if err != nil {
 		return fmt.Errorf("prepare DELETE FROM invite %v to %v, statement failed: %v", fromEmail, toEmail, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(fromEmail, toEmail)
 	if err != nil {
@@ -82,11 +81,11 @@ func DeleteInvite(fromEmail string, toEmail string) error {
 }
 
 func InsertInvite(id uuid.UUID, fromEmail string, toEmail string, now time.Time) error {
-	stmt, err := dbLib.DB.Prepare("INSERT INTO invite (id, from_email, to_email, created_at) VALUES ($1, $2, $3, $4)")
+	stmt, err := DB.Prepare("INSERT INTO invite (id, from_email, to_email, created_at) VALUES ($1, $2, $3, $4)")
 	if err != nil {
 		return fmt.Errorf("prepare INSERT INTO invite for %v to %v, statement failed: %v", fromEmail, toEmail, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(id, fromEmail, toEmail, now)
 	if err != nil {
@@ -96,11 +95,11 @@ func InsertInvite(id uuid.UUID, fromEmail string, toEmail string, now time.Time)
 }
 
 func UpdateConfirmInviteAt(fromEmail string, toEmail string, now time.Time) error {
-	stmt, err := dbLib.DB.Prepare("UPDATE invite SET confirmed_at = $1 WHERE from_email = $2 and to_email=$3")
+	stmt, err := DB.Prepare("UPDATE invite SET confirmed_at = $1 WHERE from_email = $2 and to_email=$3")
 	if err != nil {
 		return fmt.Errorf("prepare UPDATE invite statement failed: %v", err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	res, err := stmt.Exec(now, fromEmail, toEmail)
 	if err != nil {
