@@ -3,7 +3,6 @@ package db
 import (
 	"database/sql"
 	"fmt"
-	dbLib "github.com/flatfeestack/go-lib/database"
 	"github.com/google/uuid"
 	"github.com/lib/pq"
 	"log/slog"
@@ -45,13 +44,13 @@ type PayTransferEvent struct {
 }
 
 func InsertPayInEvent(payInEvent PayInEvent) error {
-	stmt, err := dbLib.DB.Prepare(`INSERT INTO payment_in_event(id, external_id, user_id, balance,  
+	stmt, err := DB.Prepare(`INSERT INTO payment_in_event(id, external_id, user_id, balance,  
                                                                 currency, status, seats, freq, created_at) 
                                    VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9)`)
 	if err != nil {
 		return fmt.Errorf("prepare INSERT INTO payment_in_event for %v statement event: %v", payInEvent.UserId, err)
 	}
-	defer dbLib.CloseAndLog(stmt)
+	defer CloseAndLog(stmt)
 
 	var res sql.Result
 	b := payInEvent.Balance.String()
@@ -67,11 +66,11 @@ func FindPayInUser(userId uuid.UUID) ([]PayInEvent, error) {
 	s := `SELECT balance, currency, status, seats, freq, created_at 
           FROM payment_in_event 
           WHERE user_id = $1`
-	rows, err := dbLib.DB.Query(s, userId)
+	rows, err := DB.Query(s, userId)
 	if err != nil {
 		return nil, err
 	}
-	defer dbLib.CloseAndLog(rows)
+	defer CloseAndLog(rows)
 
 	payInEvents := []PayInEvent{}
 	for rows.Next() {
@@ -92,7 +91,7 @@ func FindPayInUser(userId uuid.UUID) ([]PayInEvent, error) {
 func FindPayInExternal(externalId uuid.UUID, status string) (*PayInEvent, error) {
 	var payInEvent PayInEvent
 	var b string
-	err := dbLib.DB.
+	err := DB.
 		QueryRow(`SELECT user_id, balance, currency, status, seats, freq, created_at 
           FROM payment_in_event 
           WHERE external_id = $1 and status = $2`, externalId, status).
@@ -118,7 +117,7 @@ func FindPayInExternal(externalId uuid.UUID, status string) (*PayInEvent, error)
 }
 
 func FindSumPaymentByCurrency(userId uuid.UUID, status string) (map[string]*big.Int, error) {
-	rows, err := dbLib.DB.
+	rows, err := DB.
 		Query(`SELECT currency, COALESCE(sum(balance), 0)
                FROM payment_in_event 
                WHERE user_id = $1 AND status = $2
@@ -127,7 +126,7 @@ func FindSumPaymentByCurrency(userId uuid.UUID, status string) (map[string]*big.
 	if err != nil {
 		return nil, err
 	}
-	defer dbLib.CloseAndLog(rows)
+	defer CloseAndLog(rows)
 
 	m := make(map[string]*big.Int)
 	for rows.Next() {
@@ -155,7 +154,7 @@ func FindLatestDailyPayment(userId uuid.UUID, currency string) (*big.Int, int64,
 	var seats int64
 	var freq int64
 	var c time.Time
-	err := dbLib.DB.
+	err := DB.
 		QueryRow(`SELECT balance, seats, freq, created_at
                FROM payment_in_event 
                WHERE user_id = $1 AND currency = $2 AND status = $3
@@ -214,7 +213,7 @@ func PaymentSuccess(externalId uuid.UUID, fee *big.Int) error {
 
 func sumTotalEarnedAmountForContributionIds(contributionIds []uuid.UUID) (*big.Int, error) {
 	var c string
-	err := dbLib.DB.
+	err := DB.
 		QueryRow(`SELECT COALESCE(SUM(balance), 0) AS c FROM daily_contribution WHERE id = ANY($1)`, pq.Array(contributionIds)).
 		Scan(&c)
 	switch err {
