@@ -1,7 +1,6 @@
 package db
 
 import (
-	"backend/pkg/util"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"testing"
@@ -29,8 +28,8 @@ func insertTestRepoGitUrl(t *testing.T, gitUrl string) *Repo {
 }
 
 func TestRepoNotFound(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	insertTestRepo(t)
 
 	r2, err := FindRepoById(uuid.New())
@@ -39,8 +38,8 @@ func TestRepoNotFound(t *testing.T) {
 }
 
 func TestRepoFound(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	r := insertTestRepo(t)
 
 	r2, err := FindRepoById(r.Id)
