@@ -1,7 +1,6 @@
 package db
 
 import (
-	"backend/pkg/util"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"testing"
@@ -16,8 +15,8 @@ var (
 )
 
 func TestSponsorTwice(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepo(t)
 
@@ -48,8 +47,8 @@ func TestSponsorTwice(t *testing.T) {
 }
 
 func TestUnSponsorTwice(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepo(t)
 
@@ -88,8 +87,8 @@ func TestUnSponsorTwice(t *testing.T) {
 }
 
 func TestUnSponsorWrong(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepo(t)
 
@@ -107,8 +106,8 @@ func TestUnSponsorWrong(t *testing.T) {
 }
 
 func TestSponsorWrongOrder(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepo(t)
 
@@ -137,8 +136,8 @@ func TestSponsorWrongOrder(t *testing.T) {
 }
 
 func TestSponsorWrongOrderActive(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepo(t)
 
@@ -177,8 +176,8 @@ func TestSponsorWrongOrderActive(t *testing.T) {
 }
 
 func TestSponsorCorrect(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepo(t)
 	r2 := insertTestRepoGitUrl(t, "git-url2")
@@ -235,8 +234,8 @@ func TestSponsorCorrect(t *testing.T) {
 }
 
 func TestTwoRepos(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepoGitUrl(t, "git-url")
 	r2 := insertTestRepoGitUrl(t, "git-url2")
@@ -268,8 +267,8 @@ func TestTwoRepos(t *testing.T) {
 }
 
 func TestSponsorsBetween(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepoGitUrl(t, "git-url")
 	r2 := insertTestRepoGitUrl(t, "git-url2")
@@ -301,8 +300,8 @@ func TestSponsorsBetween(t *testing.T) {
 }
 
 func TestSponsorsBetween2(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepoGitUrl(t, "git-url")
 	r2 := insertTestRepoGitUrl(t, "git-url2")
@@ -334,8 +333,8 @@ func TestSponsorsBetween2(t *testing.T) {
 }
 
 func TestSponsorsBetween3(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepoGitUrl(t, "git-url")
 	r2 := insertTestRepoGitUrl(t, "git-url2")
@@ -377,8 +376,8 @@ func TestSponsorsBetween3(t *testing.T) {
 }
 
 func TestSponsorsBetween4(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	r := insertTestRepoGitUrl(t, "git-url")
 	r2 := insertTestRepoGitUrl(t, "git-url2")
