@@ -1,7 +1,6 @@
 package db
 
 import (
-	"backend/pkg/util"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"testing"
@@ -25,8 +24,8 @@ func insertTestUser(t *testing.T, email string) *UserDetail {
 }
 
 func TestUserNotFound(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	insertTestUser(t, "email")
 
 	u2, err := FindUserByEmail("email2")
@@ -35,8 +34,8 @@ func TestUserNotFound(t *testing.T) {
 }
 
 func TestUserFound(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	insertTestUser(t, "email")
 
 	u3, err := FindUserByEmail("email")
@@ -45,8 +44,8 @@ func TestUserFound(t *testing.T) {
 }
 
 func TestUserUpdate(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 
 	u.Email = "email2"
@@ -66,8 +65,8 @@ func TestUserUpdate(t *testing.T) {
 }
 
 func TestUserUpdateSeat(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	UpdateSeatsFreq(u.Id, 12, 13)
 	u2, err := FindUserById(u.Id)
@@ -76,8 +75,8 @@ func TestUserUpdateSeat(t *testing.T) {
 }
 
 func TestUserUpdateInvite(t *testing.T) {
-	util.SetupTestData()
-	defer util.TeardownTestData()
+	SetupTestData()
+	defer TeardownTestData()
 	u := insertTestUser(t, "email")
 	i := uuid.New()
 	UpdateUserInviteId(u.Id, i)
