@@ -2,12 +2,10 @@ package main
 
 import (
 	"bytes"
-	"context"
 	"crypto/sha256"
 	"encoding/base64"
 	"encoding/json"
 	"flag"
-	dbLib "github.com/flatfeestack/go-lib/database"
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -15,6 +13,7 @@ import (
 	"io"
 	"log"
 	"net/http"
+	"net/http/httptest"
 	"net/url"
 	"os"
 	"strconv"
@@ -26,7 +25,7 @@ import (
 const (
 	testDBPath    = "/tmp/fa.db"
 	testDBDriver  = "sqlite3"
-	testDBScripts = "rmdbLib.DB.sql:init.sql"
+	testDBScripts = "rmdb.sql:init.sql"
 	testDomain    = "localhost"
 	testPort      = 8082
 	testUrl       = "http://" + testDomain + ":8082"
@@ -46,45 +45,36 @@ var (
 curl -v "http://localhost:8080/signup"   -X POST   -d "{\"email\":\"tom\",\"password\":\"test\"}"   -H "Content-Type: application/json"
 */
 func TestSignup(t *testing.T) {
-	shutdown := mainTest(testParams...)
-	resp := doSignup("tom@test.ch", "testtest")
+	router := mainTest(testParams...)
+	resp := doSignup(router, "tom@test.ch", "testtest")
 
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestSignupWrongEmail(t *testing.T) {
-	shutdown := mainTest(testParams...)
-	resp := doSignup("tomtest.ch", "testtest")
+	router := mainTest(testParams...)
+	resp := doSignup(router, "tomtest.ch", "testtest")
 
 	assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
 	bodyBytes, _ := io.ReadAll(resp.Body)
 	bodyString := string(bodyBytes)
 	assert.True(t, strings.Index(bodyString, "Oops something went wrong. Please try again.") > 0)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestSignupTwiceWorking(t *testing.T) {
-	shutdown := mainTest(testParams...)
-	resp := doSignup("tom@test.ch", "testtest")
-	resp = doSignup("tom@test.ch", "testtest")
+	router := mainTest(testParams...)
+	resp := doSignup(router, "tom@test.ch", "testtest")
+	resp = doSignup(router, "tom@test.ch", "testtest")
 
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestSignupTwiceNotWorking(t *testing.T) {
-	shutdown := mainTest(testParams...)
-	resp := doSignup("tom@test.ch", "testtest")
+	router := mainTest(testParams...)
+	resp := doSignup(router, "tom@test.ch", "testtest")
 	token := token("tom@test.ch")
-	resp = doConfirm("tom@test.ch", token)
-	resp = doSignup("tom@test.ch", "testtest")
+	resp = doConfirm(router, "tom@test.ch", token)
+	resp = doSignup(router, "tom@test.ch", "testtest")
 
 	bodyBytes, _ := io.ReadAll(resp.Body)
 	bodyString := string(bodyBytes)
@@ -92,124 +82,105 @@ func TestSignupTwiceNotWorking(t *testing.T) {
 	assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
 	log.Println(bodyString)
 	assert.True(t, strings.Index(bodyString, "Oops something went wrong. Please try again.") > 0)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestConfirm(t *testing.T) {
-	shutdown := mainTest(testParams...)
-	resp := doSignup("tom@test.ch", "testtest")
+	router := mainTest(testParams...)
+	resp := doSignup(router, "tom@test.ch", "testtest")
 	assert.Equal(t, 200, resp.StatusCode)
 
 	token := token("tom@test.ch")
-	resp = doConfirm("tom@test.ch", token)
+	resp = doConfirm(router, "tom@test.ch", token)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestLogin(t *testing.T) {
-	shutdown := mainTest(testParams...)
-	resp := doSignup("tom@test.ch", "testtest")
+	router := mainTest(testParams...)
+	resp := doSignup(router, "tom@test.ch", "testtest")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
 	token := token("tom@test.ch")
-	resp = doConfirm("tom@test.ch", token)
+	resp = doConfirm(router, "tom@test.ch", token)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
-	resp = doLogin("tom@test.ch", "testtest", "", "")
+	resp = doLogin(router, "tom@test.ch", "testtest", "", "")
 	assert.Equal(t, http.StatusSeeOther, resp.StatusCode)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestLoginFalse(t *testing.T) {
-	shutdown := mainTest(testParams...)
-	resp := doAll("tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
+	router := mainTest(testParams...)
+	resp := doAll(router, "tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
-	resp = doLogin("tom@test.ch", "testtest", "", "0123456789012345678901234567890123456789012")
+	resp = doLogin(router, "tom@test.ch", "testtest", "", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusSeeOther, resp.StatusCode)
 
-	resp = doLogin("tom@test.ch", "testtest2", "", "0123456789012345678901234567890123456789012")
+	resp = doLogin(router, "tom@test.ch", "testtest2", "", "0123456789012345678901234567890123456789012")
 	bodyBytes, _ := io.ReadAll(resp.Body)
 	assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
 	assert.True(t, strings.Index(string(bodyBytes), "Oops something went wrong. Please try again.") > 0)
 
-	resp = doLogin("tom@test.ch2", "testtest", "", "0123456789012345678901234567890123456789012")
+	resp = doLogin(router, "tom@test.ch2", "testtest", "", "0123456789012345678901234567890123456789012")
 	bodyBytes, _ = io.ReadAll(resp.Body)
 	assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
 	assert.True(t, strings.Index(string(bodyBytes), "Oops something went wrong. Please try again.") > 0)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestRefresh(t *testing.T) {
 	tmp := append(testParams, "-expire-refresh=10")
-	shutdown := mainTest(tmp...)
-	resp := doAll("tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
+	router := mainTest(tmp...)
+	resp := doAll(router, "tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 	oauth := OAuth{}
 	json.NewDecoder(resp.Body).Decode(&oauth)
 	assert.NotEqual(t, "", oauth.AccessToken)
-	shutdown()
 }
 
 func TestReset(t *testing.T) {
 	tmp := append(testParams, "-expire-refresh=1")
-	shutdown := mainTest(tmp...)
+	router := mainTest(tmp...)
 
-	resp := doAll("tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
+	resp := doAll(router, "tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
-	resp = doReset("tom@test.ch")
+	resp = doReset(router, "tom@test.ch")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
 	token, _ := getForgotEmailToken("tom@test.ch")
 
-	resp = doConfirmReset("tom@test.ch", token, "testtest2")
+	resp = doConfirmReset(router, "tom@test.ch", token, "testtest2")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
-	resp = doLogin("tom@test.ch", "testtest2", "", "0123456789012345678901234567890123456789012")
+	resp = doLogin(router, "tom@test.ch", "testtest2", "", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusSeeOther, resp.StatusCode)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestResetFailed(t *testing.T) {
 	tmp := append(testParams, "-expire-refresh=1")
-	shutdown := mainTest(tmp...)
-	resp := doAll("tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
+	router := mainTest(tmp...)
+	resp := doAll(router, "tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
-	resp = doReset("tom@test.ch")
+	resp = doReset(router, "tom@test.ch")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
 	token, _ := getForgotEmailToken("tom@test.ch")
 
-	resp = doConfirmReset("tom@test.ch", token, "testtest2")
+	resp = doConfirmReset(router, "tom@test.ch", token, "testtest2")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
-	resp = doLogin("tom@test.ch", "testtest", "", "0123456789012345678901234567890123456789012")
+	resp = doLogin(router, "tom@test.ch", "testtest", "", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestTOTP(t *testing.T) {
-	shutdown := mainTest(testParams...)
-	resp := doAll("tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
+	router := mainTest(testParams...)
+	resp := doAll(router, "tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 	oauth := OAuth{}
 	json.NewDecoder(resp.Body).Decode(&oauth)
 
-	resp = doTOTP(oauth.AccessToken)
+	resp = doTOTP(router, oauth.AccessToken)
 	p := ProvisioningUri{}
 	bodyBytes, _ := io.ReadAll(resp.Body)
 	json.Unmarshal(bodyBytes, &p)
@@ -223,51 +194,47 @@ func TestTOTP(t *testing.T) {
 	totp := newTOTP(secret)
 	conf := totp.Now()
 
-	resp = doTOTPConfirm(conf, oauth.AccessToken)
+	resp = doTOTPConfirm(router, conf, oauth.AccessToken)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
-
-	resp.Body.Close()
-	shutdown()
 }
 
 func TestLoginTOTP(t *testing.T) {
-	shutdown := mainTest(testParams...)
-	resp := doAll("tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
+	router := mainTest(testParams...)
+	resp := doAll(router, "tom@test.ch", "testtest", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
 	oauth := OAuth{}
 	json.NewDecoder(resp.Body).Decode(&oauth)
 
-	totp := doAllTOTP(oauth.AccessToken)
+	totp := doAllTOTP(router, oauth.AccessToken)
 
-	resp = doLogin("tom@test.ch", "testtest", totp.Now(), "0123456789012345678901234567890123456789012")
+	resp = doLogin(router, "tom@test.ch", "testtest", totp.Now(), "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusSeeOther, resp.StatusCode)
 
-	resp = doLogin("tom@test.ch", "testtest", "", "0123456789012345678901234567890123456789012")
+	resp = doLogin(router, "tom@test.ch", "testtest", "", "0123456789012345678901234567890123456789012")
 	assert.Equal(t, http.StatusForbidden, resp.StatusCode)
-
-	resp.Body.Close()
-	shutdown()
 }
 
-func doTOTP(token string) *http.Response {
+func doTOTP(router *http.ServeMux, token string) *http.Response {
 	req, _ := http.NewRequest("POST", testUrl+"/setup/totp", nil)
 	req.Header.Set("Content-Type", "application/json")
 	req.Header.Set("Authorization", "Bearer "+token)
-	resp, _ := http.DefaultClient.Do(req)
-	return resp
+	rr := httptest.NewRecorder()
+	router.ServeHTTP(rr, req)
+	return rr.Result()
 }
 
-func doTOTPConfirm(conf string, token string) *http.Response {
+func doTOTPConfirm(router *http.ServeMux, conf string, token string) *http.Response {
 	req, _ := http.NewRequest("POST", testUrl+"/confirm/totp/"+conf, nil)
 	req.Header.Set("Content-Type", "application/json")
 	req.Header.Set("Authorization", "Bearer "+token)
-	resp, _ := http.DefaultClient.Do(req)
-	return resp
+	rr := httptest.NewRecorder()
+	router.ServeHTTP(rr, req)
+	return rr.Result()
 }
 
-func doAllTOTP(token string) *gotp.TOTP {
-	resp := doTOTP(token)
+func doAllTOTP(router *http.ServeMux, token string) *gotp.TOTP {
+	resp := doTOTP(router, token)
 	p := ProvisioningUri{}
 	bodyBytes, _ := io.ReadAll(resp.Body)
 	json.Unmarshal(bodyBytes, &p)
@@ -280,11 +247,11 @@ func doAllTOTP(token string) *gotp.TOTP {
 	secret := u.Query().Get("secret")
 	totp := newTOTP(secret)
 	conf := totp.Now()
-	resp = doTOTPConfirm(conf, token)
+	resp = doTOTPConfirm(router, conf, token)
 	return totp
 }
 
-func doConfirmReset(email string, token string, password string) *http.Response {
+func doConfirmReset(router *http.ServeMux, email string, token string, password string) *http.Response {
 	data := Credentials{
 		Email:      email,
 		Password:   password,
@@ -294,48 +261,54 @@ func doConfirmReset(email string, token string, password string) *http.Response 
 	body := bytes.NewReader(payloadBytes)
 	req, _ := http.NewRequest("POST", testUrl+"/confirm/reset", body)
 	req.Header.Set("Content-Type", "application/json")
-	resp, _ := http.DefaultClient.Do(req)
-	return resp
+	rr := httptest.NewRecorder()
+	router.ServeHTTP(rr, req)
+	return rr.Result()
 }
 
-func doReset(email string) *http.Response {
+func doReset(router *http.ServeMux, email string) *http.Response {
 	req, _ := http.NewRequest("POST", testUrl+"/reset/"+email, nil)
-	resp, _ := http.DefaultClient.Do(req)
-	return resp
+	rr := httptest.NewRecorder()
+	router.ServeHTTP(rr, req)
+	return rr.Result()
 }
 
-func doAll(email string, pass string, secret string) *http.Response {
-	resp := doSignup(email, pass)
+func doAll(router *http.ServeMux, email string, pass string, secret string) *http.Response {
+	resp := doSignup(router, email, pass)
 	token := token(email)
-	resp = doConfirm(email, token)
-	resp = doLogin(email, pass, "", secret)
+	resp = doConfirm(router, email, token)
+	resp = doLogin(router, email, pass, "", secret)
 	code := resp.Header.Get("Location")[6:]
-	resp = doCode(code, secret)
+	resp = doCode(router, code, secret)
 	return resp
 }
 
-func doCode(codeToken string, codeVerifier string) *http.Response {
+func doCode(router *http.ServeMux, codeToken string, codeVerifier string) *http.Response {
 	data := url.Values{}
 	data.Set("grant_type", "authorization_code")
 	data.Set("code", codeToken)
 	data.Set("code_verifier", codeVerifier)
 	req, _ := http.NewRequest("POST", testUrl+"/oauth/token", strings.NewReader(data.Encode()))
 	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	resp, _ := http.DefaultClient.Do(req)
-	return resp
+
+	rr := httptest.NewRecorder()
+	router.ServeHTTP(rr, req)
+	return rr.Result()
 }
 
-func doRefresh(refreshToken string) *http.Response {
+func doRefresh(router *http.ServeMux, refreshToken string) *http.Response {
 	data := url.Values{}
 	data.Set("grant_type", "refresh_token")
 	data.Set("refresh_token", refreshToken)
 	req, _ := http.NewRequest("POST", testUrl+"/oauth/token", strings.NewReader(data.Encode()))
 	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	resp, _ := http.DefaultClient.Do(req)
-	return resp
+
+	rr := httptest.NewRecorder()
+	router.ServeHTTP(rr, req)
+	return rr.Result()
 }
 
-func doLogin(email string, pass string, totp string, secret string) *http.Response {
+func doLogin(router *http.ServeMux, email string, pass string, totp string, secret string) *http.Response {
 	h := sha256.Sum256([]byte(secret))
 	data := Credentials{
 		Email:                   email,
@@ -345,22 +318,17 @@ func doLogin(email string, pass string, totp string, secret string) *http.Respon
 		CodeCodeChallengeMethod: "S256",
 	}
 
-	//do not follow redirects: https://stackoverflow.com/questions/23297520/how-can-i-make-the-go-http-client-not-follow-redirects-automatically
-	client := &http.Client{
-		CheckRedirect: func(req *http.Request, via []*http.Request) error {
-			return http.ErrUseLastResponse
-		},
-	}
-
 	payloadBytes, _ := json.Marshal(data)
 	body := bytes.NewReader(payloadBytes)
 	req, _ := http.NewRequest(http.MethodPost, testUrl+"/login", body)
 	req.Header.Set("Content-Type", "application/json")
-	resp, _ := client.Do(req)
-	return resp
+
+	rr := httptest.NewRecorder()
+	router.ServeHTTP(rr, req)
+	return rr.Result()
 }
 
-func doSignup(email string, pass string) *http.Response {
+func doSignup(router *http.ServeMux, email string, pass string) *http.Response {
 	data := Credentials{
 		Email:    email,
 		Password: pass,
@@ -373,29 +341,32 @@ func doSignup(email string, pass string) *http.Response {
 		log.Printf("request failed %v", err)
 	}
 	req.Header.Set("Content-Type", "application/json")
-	resp, err := http.DefaultClient.Do(req)
+
+	rr := httptest.NewRecorder()
+	router.ServeHTTP(rr, req)
+	return rr.Result()
+}
+
+func doConfirm(router *http.ServeMux, email string, token string) *http.Response {
+
+	req, err := http.NewRequest("GET", testUrl+"/confirm/signup/"+email+"/"+token, nil)
 	if err != nil {
 		log.Printf("request failed %v", err)
 	}
-	return resp
-}
 
-func doConfirm(email string, token string) *http.Response {
-	c := &http.Client{
-		Timeout: 15 * time.Second,
-	}
-	resp, _ := c.Get(testUrl + "/confirm/signup/" + email + "/" + token)
-	return resp
+	rr := httptest.NewRecorder()
+	router.ServeHTTP(rr, req)
+	return rr.Result()
 }
 
 func token(email string) string {
 	r, _ := getEmailToken(email)
-	return string(r)
+	return r
 }
 
 func getEmailToken(email string) (string, error) {
 	var emailToken string
-	err := dbLib.DB.QueryRow("SELECT email_token from auth where email = $1", email).Scan(&emailToken)
+	err := DB.QueryRow("SELECT email_token from auth where email = $1", email).Scan(&emailToken)
 	if err != nil {
 		return "", err
 	}
@@ -404,7 +375,7 @@ func getEmailToken(email string) (string, error) {
 
 func getForgotEmailToken(email string) (string, error) {
 	var forgetEmailToken string
-	err := dbLib.DB.QueryRow("SELECT forget_email_token from auth where email = $1", email).Scan(&forgetEmailToken)
+	err := DB.QueryRow("SELECT forget_email_token from auth where email = $1", email).Scan(&forgetEmailToken)
 	if err != nil {
 		return "", err
 	}
@@ -417,45 +388,21 @@ func TestSecret(t *testing.T) {
 	assert.Equal(t, "n4bQgYhMfWWaL-qgxVrQFaO_TxsrC4Is0V1sFbDwCgg", s)
 }
 
-func mainTest(args ...string) func() {
+func mainTest(args ...string) *http.ServeMux {
 	oldArgs := os.Args
 	os.Args = []string{oldArgs[0]}
 	os.Args = append(os.Args, args...)
+	flag.CommandLine = flag.NewFlagSet(os.Args[0], flag.ExitOnError) //flags are now reset
 
 	opts = NewOpts()
 	var err error
-	err = dbLib.InitDb(opts.DBDriver, opts.DBPath, opts.DBScripts)
+	err = InitDb(opts.DBDriver, opts.DBPath, opts.DBScripts)
 	if err != nil {
 		log.Fatal(err)
 	}
 	addInitialUsers()
-	serverRest, doneChannelRest, err := serverRest(false)
-	if err != nil {
-		log.Fatal(err)
-	}
 
-	return func() {
-		defer timeTrack(time.Now(), "shutdown")
-		os.Args = oldArgs
-		flag.CommandLine = flag.NewFlagSet(os.Args[0], flag.ExitOnError)
-
-		serverRest.Shutdown(context.Background())
-
-		err := dbLib.DB.Close()
-		if err != nil {
-			log.Printf("could not close DB %v", err)
-		}
-
-		//for testing, the DB needs to be wiped after each run
-		if opts.DBDriver == "sqlite3" {
-			err = os.Remove(opts.DBPath)
-			if err != nil {
-				log.Printf("could not remove DB file %v", err)
-			}
-		}
-
-		<-doneChannelRest
-	}
+	return setupRouter()
 }
 
 func timeTrack(start time.Time, name string) {
